# backend/openapi.yml
openapi: 3.0.3
info:
  title: CapeConnect API
  version: 1.1.0
  description: >
    Live backend contract for the current CapeConnect static frontend and
    supporting admin tools. Authentication uses opaque bearer session tokens,
    not JWTs.
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    Wallet:
      type: object
      properties:
        user_id:
          type: string
        balance_cents:
          type: integer
        currency:
          type: string
        operator:
          type: string
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        type:
          type: string
        amount_cents:
          type: integer
        ref_ticket_id:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/wallets/me:
    get:
      tags: [Wallets]
      summary: Get current wallet and recent transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallet:
                    $ref: '#/components/schemas/Wallet'
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/WalletTransaction'
  /api/wallets/topup:
    post:
      tags: [Wallets]
      summary: Directly top up the wallet balance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountCents]
              properties:
                amountCents:
                  type: integer
                operator:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Updated wallet
  /api/wallets/spend:
    post:
      tags: [Wallets]
      summary: Debit wallet balance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountCents]
              properties:
                amountCents:
                  type: integer
                operator:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Updated wallet
        '400':
          description: Insufficient balance
...